requirements.txt
asyncpraw
python-telegram-bot
python-telegram-bot[job-queue]
aiohttp
python-dotenv
opencv-python
pillow
redgifs